+++ Include/RmlUi/Core/Elements/ElementProgressBar.h
@@ -67,10 +67,13 @@ public:
 	ElementProgressBar(const String& tag);
 	virtual ~ElementProgressBar();
 
-	/// Return the value of the progress bar [0, 1]
+	/// Returns the value of the progress bar.
 	float GetValue() const;
 
-	/// Set the value of the progress bar
+	/// Returns the maximum value of the progress bar.
+	float GetMax() const;
+
+	/// Sets the value of the progress bar.
 	void SetValue(float value);
 
 	/// Returns the element's inherent size.
@@ -97,8 +100,6 @@ private:
 
 	Direction direction;
 	StartEdge start_edge;
-
-	float value;
 
 	Element* fill;